+++ core/triggers/interface_90_modSubtotal_subtotaltrigger.class.php
@@ -346,7 +346,6 @@ class Interfacesubtotaltrigger extends DolibarrTriggers
 								$desc = $this->getShippingList($commande->id);
 							}
 
-
                             if(!empty($current_fk_commande)) {
                                 $subtotal_skip = true;
                                 TSubtotal::addTitle($facture, $label, 1, $rang, $desc);
@@ -363,7 +362,7 @@ class Interfacesubtotaltrigger extends DolibarrTriggers
 					    {
                             $subtotal_skip = true;
                             $subtotal_bloc_already_add_st = 1;
-							$rang+=2; // pour eviter un bug de décalage ou le sous total ce retrouve apres le nouveau titre : dug constaté en V16 ne doit pas avoir d'impact sur les anciennes versions
+							$rang += 2; // pour eviter un bug de décalage ou le sous total ce retrouve apres le nouveau titre : dug constaté en V16 ne doit pas avoir d'impact sur les anciennes versions
                             TSubtotal::addTotal($facture, $langs->trans('SubTotal'), 1, $rang);
                             $subtotal_bloc_already_add_st = 0;
                             $rang++;